backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pydantic-settings==2.1.0
PyPDF2==3.0.1
python-docx==0.8.11
